Switch to f32

+++ src/cameras.h
@@ -9,11 +9,11 @@
 #endif
 
 struct ImagePlane {
-	f64 fov;
-	f64 aspect_ratio;
-	f64 pixel_size;
-	f64 height;
-	f64 width;
+	f32 fov;
+	f32 aspect_ratio;
+	f32 pixel_size;
+	f32 height;
+	f32 width;
 	u32 rows;
 	u32 cols;
 };
@@ -23,18 +23,18 @@ struct Camera {
 	Vec3D normal;
 	Vec3D up;
 	ImagePlane image_plane;
-	f64 aperture;
-	f64 focal_distance;
+	f32 aperture;
+	f32 focal_distance;
 };
 
-ImagePlane create_image_plane(f64 fov, f64 aspect_ratio, u32 pixel_height) {
+ImagePlane create_image_plane(f32 fov, f32 aspect_ratio, u32 pixel_height) {
 	ImagePlane image_plane = {};
-	f64 fov_radians = fov * (M_PI / 180.0);
+	f32 fov_radians = fov * (M_PI / 180.0);
 	image_plane.fov = fov;
 	image_plane.aspect_ratio = aspect_ratio;
 	image_plane.height = 2.0 * std::tan(fov_radians / 2.0);
 	image_plane.width = aspect_ratio * image_plane.height;
-	image_plane.pixel_size = image_plane.height / (f64) pixel_height;
+	image_plane.pixel_size = image_plane.height / (f32) pixel_height;
 	image_plane.rows = (u32) pixel_height;
 	image_plane.cols = (u32) std::round(image_plane.width / image_plane.pixel_size);
 	return image_plane;
